refactor: add slices for columns and profile
