+++ src/pages/Board/Board.tsx
@@ -8,7 +8,8 @@ import s from './Board.module.scss';
 
 const Board: FC = () => {
   const { selectBoard, reloadStatus } = useAppSelector((state) => state.boardsSlice);
-  let requestStatus;
+
+  const { reloadColumnsStatus } = useAppSelector((state) => state.columnsSlice);
   const { title, description, columns } = selectBoard;
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
@@ -18,10 +19,10 @@ const Board: FC = () => {
   useEffect(() => {
     const id = localStorage.getItem('selectBoard');
 
-    if (id && reloadStatus) {
+    if (id && reloadColumnsStatus) {
       dispatch(getBoardById(id));
     }
-  }, [dispatch, reloadStatus]);
+  }, [dispatch, reloadColumnsStatus]);
 
   return (
     <section className={s.boardPage}>
@@ -29,7 +30,7 @@ const Board: FC = () => {
         <button className={s.homeBtn} onClick={toMain}></button>
         <p className={s.boardTitle}>{title}</p>
         <p className={s.text}>{description}</p>
-        {requestStatus === 'pending' ? <div>Loading</div> : <ColumnsField columns={columns} />}
+        {reloadStatus ? <div>Loading</div> : <ColumnsField columns={columns} />}
       </div>
     </section>
   );
